requirements.txt
streamlit
pandas
numpy
matplotlib
fpdf2
xlsxwriter
python-dateutil